requirements.txt
python-telegram-bot==13.15
environs==9.5.0
google-cloud-dialogflow==2.23.3
google-cloud-api-keys==0.5.3

